functions/requirements.txt

google-cloud-firestore
google-cloud-storage
requests
google-cloud-secret-manager